+++ packages/core/src/lib/list/list-config.interface.ts
@@ -56,4 +56,6 @@ export interface ListConfig<T> extends ItemConfig<T> {
   storageKey?: string | ((list: List<T>) => string);
   /** Transforms the Items before they are displayed, e.g. to apply a filter for the view **/
   display?: (items: Item<T>[]) => Item<T>[];
+  /** If set, a filter input for the given field property will be shown by default */
+  defaultFilter?: string | boolean;
 }