+++ packages/data/src/lib/entry-list-pop/entry-list-pop.component.ts
@@ -44,7 +44,7 @@ export class EntryListPopComponent extends PopComponent implements OnChanges {
         } else if (this.selection) {
             this.selection.toggle(item);
         }
-        this.searchbar.focusEvent.emit(true);
+        // this.searchbar.focusEvent.emit(true);
     }
 
     getHeader(entryList) {